requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.2
djangorestframework==3.14.0
environ==1.0
pytz==2023.3
sentry-sdk==1.28.1
sqlparse==0.4.4
urllib3==2.0.4
